çalışmalar devam ediyor.

+++ Library/Circle.Library.Business/Handlers/GroupClaims/Commands/DeleteGroupClaimCommand.cs
@@ -9,33 +9,38 @@ using Circle.Core.Utilities.Results;
 using Circle.Library.DataAccess.Abstract;
 using MediatR;
 using System;
+using Circle.Core.Utilities.Messages;
+using Circle.Library.Business.Helpers;
 
 namespace Circle.Library.Business.Handlers.GroupClaims.Commands
 {
-    public class DeleteGroupClaimCommand : IRequest<IResult>
+    public class DeleteGroupClaimCommand : IRequest<ResponseMessage<NoContent>>
     {
         public Guid Id { get; set; }
 
-        public class DeleteGroupClaimCommandHandler : IRequestHandler<DeleteGroupClaimCommand, IResult>
+        public class DeleteGroupClaimCommandHandler : IRequestHandler<DeleteGroupClaimCommand, ResponseMessage<NoContent>>
         {
             private readonly IGroupClaimRepository _groupClaimRepository;
+            private readonly IReturnUtility _returnUtility;
 
-            public DeleteGroupClaimCommandHandler(IGroupClaimRepository groupClaimRepository)
+            public DeleteGroupClaimCommandHandler(IGroupClaimRepository groupClaimRepository, IReturnUtility returnUtility)
             {
                 _groupClaimRepository = groupClaimRepository;
+                _returnUtility = returnUtility;
             }
 
-            [SecuredOperation(Priority = 1)]
-            [CacheRemoveAspect("Get")]
-            [LogAspect(typeof(MsSqlLogger))]
-            public async Task<IResult> Handle(DeleteGroupClaimCommand request, CancellationToken cancellationToken)
+            public async Task<ResponseMessage<NoContent>> Handle(DeleteGroupClaimCommand request, CancellationToken cancellationToken)
             {
-                var groupClaimToDelete = await _groupClaimRepository.GetAsync(x => x.GroupId == request.Id);
+                var groupClaimToDelete = await _groupClaimRepository.GetAsync(x => x.Id == request.Id);
+                if (groupClaimToDelete == null || groupClaimToDelete.Id == Guid.Empty)
+                {
+                    return await _returnUtility.NoDataFound<NoContent>(MessageDefinitions.KAYIT_BULUNAMADI);
+                }
 
                 _groupClaimRepository.Delete(groupClaimToDelete);
                 await _groupClaimRepository.SaveChangesAsync();
 
-                return new SuccessResult(null);
+                return await _returnUtility.Success<NoContent>(MessageDefinitions.SILME_ISLEMI_BASARILI);
             }
         }
     }
